#13 - Refactor : hashTag -> postHashTag 패키지명 수정
