+++ 1Week_Mission/mutbooks/src/main/java/com/example/mutbooks/app/post/repository/PostRepositoryImpl.java
@@ -8,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 
 import java.util.List;
 
-import static com.example.mutbooks.app.hashTag.entity.QPostHashTag.postHashTag;
+import static com.example.mutbooks.app.postHashTag.entity.QPostHashTag.postHashTag;
 import static com.example.mutbooks.app.keyword.entity.QPostKeyword.postKeyword;
 import static com.example.mutbooks.app.post.entity.QPost.post;
 
@@ -18,7 +18,7 @@ public class PostRepositoryImpl implements PostRepositoryCustom{
 
     @Override
     public List<Post> searchQsl(Member author, String kwType, String kw) {
-        // 1. 전체 조회(키워드가 없거나 검색타입이 검색 타입이 hashTag 가 아닌 경우)
+        // 1. 전체 조회(키워드가 없거나 검색타입이 검색 타입이 postHashTag 가 아닌 경우)
         JPAQuery<Post> jpqQuery = jpaQueryFactory
                 .select(post)
                 .where(post.author.id.eq(author.getId()))
@@ -27,7 +27,7 @@ public class PostRepositoryImpl implements PostRepositoryCustom{
 
         // 2. 키워드 기반 검색(키워드가 존재하고 키워드 타입이 해시태그인 경우)
         if(!kw.equals("")) {
-            if(kwType.equals("hashTag")) {
+            if(kwType.equals("postHashTag")) {
                 jpqQuery
                         .innerJoin(postHashTag)
                         .on(post.eq(postHashTag.post))
